# music-player-backend/src/main/resources/api.yml
openapi: 3.1.1
info:
  title: Peer-Assisted Music Streaming API
  description: Backend API for hybrid prefix caching, predictive pre-fetching, and master fallback.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /songs:
    get:
      summary: Get songs
      description: |
        Retrieves paginated song metadata.

        Visibility rule:
        - Authenticated callers receive streamable songs plus user-uploaded songs they own.
      operationId: getAllSongs
      tags:
        - Songs
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search in song name, artist name, and album name.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: name,asc
          description: Sort as `field,direction` (e.g. `name,asc`, `year,desc`).
      responses:
        '200':
          description: A page of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
    post:
      summary: Upload a new song
      description: Ingests a raw audio file (MP3/FLAC), deduplicates chunks, and stores metadata.
      operationId: uploadSong
      tags:
        - Songs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SongUploadDto'

      responses:
        '201':
          description: Song created and chunks processed
        '400':
          description: Invalid input
  /songs/negotiate:
    post:
      summary: Negotiate a user upload
      description: Client sends metadata and chunk hashes. Server returns ID and list of missing chunk indices.
      operationId: negotiateUserUpload
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NegotiationRequestDto'
      responses:
        '200':
          description: Negotiation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NegotiationResponseDto'
  /songs/recommendations:
    get:
      summary: Get personalised song recommendations
      description: |
        Paginated recommendations ordered by play count descending. Songs the
        user already played in the last 48 hours are excluded so the list
        stays fresh.
      operationId: getRecommendations
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of recommended songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
  /songs/forgotten:
    get:
      summary: Get forgotten favourites
      description: |
        Paginated list of songs the user played at least once but has not
        listened to in the last 30 days, ordered by last-played ascending so
        the most-forgotten tracks surface first.
      operationId: getForgottenFavourites
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of forgotten favourite songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
  /songs/quick-dial:
    get:
      summary: Get quick-dial songs
      description: |
        Paginated list ordered by recency, with songs that have never been
        played falling back to their added-at timestamp.
      operationId: getQuickDial
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of quick-dial songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
  /songs/favourites:
    get:
      summary: Get the user's favourite songs
      description: Paginated list of songs the user has liked, ordered by play count descending.
      operationId: getFavourites
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of liked songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
  /songs/most-played:
    get:
      summary: Get the user's most-played songs
      description: Paginated list of songs the user has played at least once, ordered by play count descending.
      operationId: getMostPlayed
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of most-played songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
  /songs/recently-played:
    get:
      summary: Get the user's recently-played songs
      description: Paginated list of songs the user has played, ordered by last-played descending.
      operationId: getRecentlyPlayed
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of recently-played songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
  /songs/{fileHash}:
    get:
      summary: Get song metadata
      description: Retrieves metadata for a specific song by file hash.
      operationId: getSongById
      tags:
        - Songs
      parameters:
        - name: fileHash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Song metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongDto'
        '404':
          description: Song not found
    patch:
      summary: Update the requesting user's library entry for a song
      description: |
        Updates per-user fields (`likedByUser`, `lastPlayed`, `playCount`) for the
        song in the current user's library. Only fields present in the request
        body are changed; missing fields stay as they were.
      operationId: updateUserSongLibrary
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: fileHash
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserSongDto'
      responses:
        '200':
          description: Updated song metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongDto'
        '401':
          description: Unauthorized
        '404':
          description: Song not found
  /songs/{fileHash}/cover:
    get:
      summary: Get song cover art
      description: Returns the binary cover image for the song's album.
      operationId: getSongCover
      tags:
        - Songs
      parameters:
        - name: fileHash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cover image bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Song or cover not found
  /songs/{fileHash}/chunks/{chunkIndex}:
    post:
      summary: Upload a missing chunk
      description: Uploads a specific 64KB binary chunk to fill a gap in a user song.
      operationId: uploadMissingChunk
      tags:
        - Songs
      security:
        - bearerAuth: [ ]
      parameters:
        - name: fileHash
          in: path
          required: true
          schema:
            type: string
        - name: chunkIndex
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - chunkData
                - contentHash
              properties:
                chunkData:
                  type: string
                  format: binary
                contentHash:
                  type: string
                  description: The SHA-256 hash of this chunk for verification.
      responses:
        '201':
          description: Chunk accepted and linked
        '400':
          description: Hash mismatch or invalid index
...
  /albums/{albumHash}/songs:
    get:
      summary: Get songs in an album
      description: Returns songs visible to the caller in the album, ordered by disc number then track number.
      operationId: getAlbumSongs
      tags:
        - Albums
      parameters:
        - name: albumHash
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of songs in the album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '404':
          description: Album not found
...
  /artists/{artistHash}/songs:
    get:
      summary: Get songs by artist
      description: Returns songs visible to the caller for this artist, ordered by song name.
      operationId: getArtistSongs
      tags:
        - Artists
      parameters:
        - name: artistHash
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of songs by the artist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '404':
          description: Artist not found
...
  /playlists/{playlistId}/songs:
    get:
      summary: Get songs in a playlist
      description: Returns playlist songs visible to the caller in persisted playlist order.
      operationId: getPlaylistSongs
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of songs in the playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Playlist not found
...
components:
  schemas:
    ArtistDto:
      type: object
      properties:
        hash:
          type: string
        name:
          type: string
...
    AlbumDto:
      type: object
      properties:
        hash:
          type: string
        name:
          type: string
...
    SongDto:
      type: object
      properties:
        fileHash:
          type: string
        name:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistDto'
        album:
          $ref: '#/components/schemas/AlbumDto'
        durationInSeconds:
          type: integer
        trackNumber:
          type: integer
        discNumber:
          type: integer
        year:
          type: integer
        lastPlayed:
          type: string
          format: date-time
          nullable: true
          description: When the requesting user last played this song; null if never.
        playCount:
          type: integer
          format: int64
          default: 0
          description: How many times the requesting user has played this song.
        likedByUser:
          type: boolean
          default: false
          description: Whether the requesting user has liked this song.
    UpdateUserSongDto:
      type: object
      description: Patch payload for the requesting user's library entry for a song.
      properties:
        likedByUser:
          type: boolean
          nullable: true
        lastPlayed:
          type: string
          format: date-time
          nullable: true
        playCount:
          type: integer
          format: int64
          nullable: true
          minimum: 0
    SongUploadDto:
      type: object
      required:
        - file
        - name
        - artistName
        - albumName
      properties:
        file:
          type: string
          format: binary
          description: The audio file (MP3, FLAC, WAV)
        name:
          type: string
        artistName:
          type: string
        albumName:
          type: string
        durationInSeconds:
          type: integer
        trackNumber:
          type: integer
        releaseYear:
          type: integer
        discNumber:
          type: integer
        photo:
          type: string
          description: Base64-encoded album art image
        fileHash:
          type: string
          description: The SHA-256 hash of the entire file for quick server-side duplicate detection.
...
    NegotiationRequestDto:
      type: object
      required:
        - name
        - hashes
      properties:
        name:
          type: string
        artistName:
          type: string
        albumName:
          type: string
        durationInSeconds:
          type: integer
        discNumber:
          type: integer
        trackNumber:
          type: integer
        releaseYear:
          type: integer
        photo:
          type: string
          description: Base64-encoded album art image
        fileHash:
          type: string
          description: The SHA-256 hash of the entire file for quick server-side duplicate detection.
        hashes:
          type: array
          description: The sequential list of SHA-256 hashes for the file chunks.
          items:
            type: string
    NegotiationResponseDto:
      type: object
      required:
        - fileHash
        - missingIndices
      properties:
        fileHash:
          type: string
        missingIndices:
          type: array
          description: The indices (0-based) of the chunks the server DOES NOT have.
          items:
            type: integer
    SongPageDto:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SongDto'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT